Implements portable mode support
Adds support for running the application in portable mode.

This allows the application to store its data in a "ModrinthAppData" folder next to the executable, instead of the standard system directories.

The portable mode is activated by either setting the `MODRINTH_PORTABLE` environment variable or by creating a `portable.txt` file in the same directory as the executable.

Environment variables are also set up so that the window state plugin respects the portable directory configuration.

Exposes a new `is_portable` tauri command.